Add HDV sell items offline mode

+++ Server/Stump.Server.WorldServer/Game/Items/BidHouse/BidHouseManager.cs
@@ -108,13 +108,18 @@ namespace Stump.Server.WorldServer.Game.Items.BidHouse
 
         public List<BidHouseItem> GetBidHouseItems(ItemTypeEnum type, int maxItemLevel)
         {
-            return m_bidHouseItems.Where(x => x.Template.TypeId == (int)type && x.Template.Level <= maxItemLevel)
+            return m_bidHouseItems.Where(x => x.Template.TypeId == (int)type && x.Template.Level <= maxItemLevel && !x.Sold)
                 .GroupBy(x => x.Template.Id).Select(x => x.First()).ToList();
+        }
+
+        public List<BidHouseItem> GetSoldBidHouseItems(int ownerId)
+        {
+            return m_bidHouseItems.Where(x => x.Record.OwnerId == ownerId && x.Sold).ToList();
         }
 
         public List<BidHouseItem> GetBidHouseItems(int ownerId, IEnumerable<int> types)
         {
-            return m_bidHouseItems.Where(x => x.Record.OwnerId == ownerId && types.Contains((int)x.Template.TypeId)).ToList();
+            return m_bidHouseItems.Where(x => x.Record.OwnerId == ownerId && types.Contains((int)x.Template.TypeId) && !x.Sold).ToList();
         }
 
         public BidHouseItem GetBidHouseItem(int guid)
@@ -122,14 +127,9 @@ namespace Stump.Server.WorldServer.Game.Items.BidHouse
             return m_bidHouseItems.FirstOrDefault(x => x.Guid == guid);
         }
 
-        public BidHouseItem GetBidHouseItem(int itemId, int quantity, int price)
-        {
-            return m_bidHouseItems.FirstOrDefault(x => x.Template.Id == itemId && x.Stack == quantity && x.Price == price);
-        }
-
         public int GetAveragePriceForItem(int itemId)
         {
-            var items = m_bidHouseItems.Where(x => x.Template.Id == itemId).Select(x => (int)(x.Price / x.Stack)).ToArray();
+            var items = m_bidHouseItems.Where(x => x.Template.Id == itemId && !x.Sold).Select(x => (int)(x.Price / x.Stack)).ToArray();
 
             if (!items.Any())
                 return 0;
@@ -168,12 +168,18 @@ namespace Stump.Server.WorldServer.Game.Items.BidHouse
 
         public event Action<BidHouseItem, BidHouseCategory, bool> ItemRemoved;
 
-        public void RemoveBidHouseItem(BidHouseItem item)
+        public void RemoveBidHouseItem(BidHouseItem item, bool removeOnly = false)
         {
-            WorldServer.Instance.IOTaskPool.AddMessage(
-                () => Database.Delete(item.Record));
+            if (!item.Sold || removeOnly)
+            {
+                WorldServer.Instance.IOTaskPool.AddMessage(
+                    () => Database.Delete(item.Record));
 
-            m_bidHouseItems.Remove(item);
+                m_bidHouseItems.Remove(item);
+            }
+
+            if (removeOnly)
+                return;
 
             var category = GetBidHouseCategory(item);
             var categoryDeleted = false;
